Factor out test setup function
